move SearchBar to components/libraries folder
